kuratorV1/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py==2.5.2",
    "email-validator==2.2.0",
    "flask==3.1.0",
    "flask-sqlalchemy==3.1.1",
    "google-api-python-client==2.166.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.1",
    "gunicorn==23.0.0",
    "pillow==11.1.0",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.1.0",
    "python-magic==0.4.27",
]
